+++ app/src/main/java/de/triology/universeadm/Constraint.java
@@ -15,7 +15,8 @@ public abstract class Constraint<T> {
   public enum ID {
     UNIQUE_USERNAME,
     UNIQUE_EMAIL,
-    UNIQUE_GROUP_NAME
+    UNIQUE_GROUP_NAME,
+    VALID_EMAIL,
   }
 
   /**